Removed a whole lot of redundant code
Consolidated all these code into a single base class that the fragments now reference off on to reduce code lines (and app size) :D
