+++ app/src/main/java/com/itachi1706/cheesecakeutilities/Modules/VehicleMileageTracker/Fragments/VehicleMileageVTypeStatsFragment.java
@@ -1,19 +1,8 @@
 package com.itachi1706.cheesecakeutilities.Modules.VehicleMileageTracker.Fragments;
 
 
-import android.content.SharedPreferences;
-import android.os.Bundle;
-import android.preference.PreferenceManager;
-import android.support.v4.app.Fragment;
 import android.support.v4.util.ArrayMap;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.widget.DefaultItemAnimator;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.util.Log;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
 import android.widget.Toast;
 
 import com.google.firebase.database.DataSnapshot;
@@ -22,15 +11,12 @@ import com.google.firebase.database.ValueEventListener;
 import com.itachi1706.cheesecakeutilities.Modules.VehicleMileageTracker.FirebaseUtils;
 import com.itachi1706.cheesecakeutilities.Modules.VehicleMileageTracker.Objects.Vehicle;
 import com.itachi1706.cheesecakeutilities.Objects.DualLineString;
-import com.itachi1706.cheesecakeutilities.R;
-import com.itachi1706.cheesecakeutilities.RecyclerAdapters.DualLineStringRecyclerAdapter;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import static com.itachi1706.cheesecakeutilities.Modules.VehicleMileageTracker.FirebaseUtils.FB_REC_STATS;
 import static com.itachi1706.cheesecakeutilities.Modules.VehicleMileageTracker.FirebaseUtils.FB_REC_USER;
-import static com.itachi1706.cheesecakeutilities.Modules.VehicleMileageTracker.FirebaseUtils.MILEAGE_DEC;
 import static com.itachi1706.cheesecakeutilities.Modules.VehicleMileageTracker.FirebaseUtils.parseData;
 
 /**
@@ -38,52 +24,10 @@ import static com.itachi1706.cheesecakeutilities.Modules.VehicleMileageTracker.F
  * for com.itachi1706.cheesecakeutilities.Modules.VehicleMileageTracker.Fragments in CheesecakeUtilities
  */
 
-public class VehicleMileageVTypeStatsFragment extends Fragment {
-
-    DualLineStringRecyclerAdapter adapter;
-    SwipeRefreshLayout refreshLayout;
-    SharedPreferences sp;
-    private boolean decimal;
+public class VehicleMileageVTypeStatsFragment extends VehicleMileageFragmentBase {
 
     private ArrayMap<String, String> vehicles;
-
-
-    public VehicleMileageVTypeStatsFragment() {
-        // Required empty public constructor
-    }
-
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        View v = inflater.inflate(R.layout.fragment_refreshable_recycler_view, container, false);
-
-        RecyclerView recyclerView = v.findViewById(R.id.recycler_view);
-        if (recyclerView != null) {
-            recyclerView.setHasFixedSize(true);
-            LinearLayoutManager linearLayoutManager = new LinearLayoutManager(getActivity());
-            linearLayoutManager.setOrientation(LinearLayoutManager.VERTICAL);
-            recyclerView.setLayoutManager(linearLayoutManager);
-            recyclerView.setItemAnimator(new DefaultItemAnimator());
-
-            // Set up blank layout for now
-            List<DualLineString> placeholder = new ArrayList<>();
-            placeholder.add(new DualLineString("Loading...", "Calculating statistics..."));
-            adapter = new DualLineStringRecyclerAdapter(placeholder, false);
-            recyclerView.setAdapter(adapter);
-
-            sp = PreferenceManager.getDefaultSharedPreferences(getActivity().getApplicationContext());
-            decimal = sp.getBoolean(MILEAGE_DEC, true);
-        }
-        refreshLayout = v.findViewById(R.id.pull_to_refresh);
-        refreshLayout.setOnRefreshListener(this::updateStats);
-        refreshLayout.setColorSchemeResources(
-                R.color.refresh_progress_1,
-                R.color.refresh_progress_2);
-
-        return v;
-    }
+    private boolean done = false;
 
     @Override
     public void onResume() {
@@ -103,6 +47,7 @@ public class VehicleMileageVTypeStatsFragment extends Fragment {
                             }
                         }
                     }
+                    done = true;
                     updateStats();
                 }
 
@@ -111,10 +56,14 @@ public class VehicleMileageVTypeStatsFragment extends Fragment {
                     Log.e("VehicleMileageStats", "Error in Firebase DB call (VType-Legend): " + databaseError.getDetails());
                 }
             });
-        } else updateStats();
+        } else {
+            done = true;
+            updateStats();
+        }
     }
 
     public void updateStats() {
+        if (!done) return;
         final String user_id = sp.getString("firebase_uid", "nien");
         if (user_id.equalsIgnoreCase("nien")) {
             // Fail, return to login activity
